Restructured directories and implemented form fill
Directories have been restructured to avoid circular imports when setting up database, using forms, and reading database. Implemented reading database values into form fields for sales database. Resolved issue with generating username
