+++ Sales_tracker/run.py
@@ -1,0 +1,5 @@
+from main import app
+from db import db
+
+db.init_app(app)
+
